M2-6502: Updated activity timer UI, code improvement (#799)
* M2-6502: Updated activity timer UI, code improvement

* M2-6502: Code review fixes
