+++ src/features/pass-survey/ui/Header.tsx
@@ -1,15 +1,11 @@
-import { useState } from 'react';
 import { StyleSheet } from 'react-native';
 
 import { CachedImage } from '@georstat/react-native-image-cache';
 import { XStack } from '@tamagui/stacks';
 import { useTranslation } from 'react-i18next';
-import DeviceInfo from 'react-native-device-info';
-import { useSafeAreaInsets } from 'react-native-safe-area-context';
 
-import { colors, HourMinute, isIphoneX } from '@shared/lib';
+import { colors } from '@shared/lib';
 import { ListSeparator, BackButton, Text, Box } from '@shared/ui';
-import TimeRemaining from '@shared/ui/TimeRemaining.tsx';
 
 import HeaderProgress from './HeaderProgress.tsx';
 
@@ -20,8 +16,6 @@ type Props = {
   flowId?: string;
   eventId: string;
   appletId: string;
-  entityStartedAt: number;
-  timer: HourMinute | null;
 };
 
 function Header({
@@ -31,43 +25,12 @@ function Header({
   flowId,
   eventId,
   appletId,
-  entityStartedAt,
-  timer,
 }: Props) {
-  const { top: safeAreaTop } = useSafeAreaInsets();
   const { t } = useTranslation();
-  const hasNotch = DeviceInfo.hasNotch();
-  const isNotIPhoneX = !isIphoneX();
-
-  const [showTimeLeft, setShowTimeLeft] = useState(!!timer);
-
-  const [timerHeight, setTimerHeight] = useState(0);
-
-  const timerMarginTop = hasNotch ? (safeAreaTop - timerHeight) / 2 : 16;
 
   return (
-    <Box>
-      {showTimeLeft && (
-        <TimeRemaining
-          mt={timerMarginTop}
-          left={16}
-          zIndex={1}
-          entityStartedAt={entityStartedAt}
-          timerSettings={timer as HourMinute}
-          clockIconShown={isNotIPhoneX}
-          opacity={timerHeight ? 1 : 0}
-          onTimeElapsed={() => setShowTimeLeft(false)}
-          onLayout={e => {
-            setTimerHeight(e.nativeEvent.layout.height);
-          }}
-        />
-      )}
-      <XStack
-        w="100%"
-        alignItems="center"
-        p={10}
-        mt={showTimeLeft ? 0 : safeAreaTop}
-      >
+    <>
+      <XStack w="100%" alignItems="center" p={10}>
         {showWatermark && watermark && (
           <Box style={styles.watermarkContainer}>
             <CachedImage
@@ -88,7 +51,7 @@ function Header({
       </XStack>
       <HeaderProgress appletId={appletId} eventId={eventId} flowId={flowId} />
       <ListSeparator mt={10} bg={colors.lighterGrey7} />
-    </Box>
+    </>
   );
 }
