+++ src/shared/ui/index.tsx
@@ -39,6 +39,7 @@ export * from './Stepper';
 export { default as Slider } from './Slider';
 export { default as Button } from './Button';
 export { default as DateTimePicker } from './DateTimePicker';
+export { default as TimeRemaining } from './TimeRemaining';
 export { default as ScrollButton } from './ScrollButton';
 export { default as ScrollableContent } from './ScrollableContent';
 export { default as TouchableOpacity } from './TouchableOpacity';
